extract into superclass
